+++ changelog.rst
@@ -2,5 +2,10 @@ Changelog
 =========
 
 
-development version: 2015-06-03
+development version: 2015-07-28
 -------------------------------
+
+* Add db.mixins with DefaultColsMixin (id, Arrow lib UTC timestamps) and MethodsMixin (incomplete).
+* Some MethodsMixin methods now have support for commit/flush parameters.
+* Add .testing:EntityBase which uses named tuples to declare the checks needed and adds some
+  additional logic.